recognition/pyproject.toml
[tool.poetry]
name = "str"
version = "0.1.0"
description = ""
authors = ["eunsung"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
uvicorn = "^0.34.0"
fastapi = "^0.115.8"
pydantic = "^2.10.6"
pyyaml = "^6.0.2"
kafka-python = "^2.0.4"
confluent-kafka = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
